backend/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
SQLAlchemy==2.0.19
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
PyMySQL==1.1.0

